Removed possibility of managing different workspaces
